Extract `DashCardActionButtons`
