+++ frontend/src/metabase/dashboard/components/DashCard/DashCardActionButtons/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DashCardActionButtons";
